自动注册支持排序 修改部分物品/方块的名称，与原版保持一致 添加青铜

+++ src/main/java/xyz/emptydreams/mi/api/tools/item/MISword.java
@@ -17,7 +17,7 @@ public class MISword extends ItemSword implements IToolMaterial {
 	private final double SPEED;
 	private final double DAMAGE;
 	
-	public MISword(ToolMaterial materialIn, float damage, float speed) {
+	public MISword(ToolMaterial materialIn, float damage, double speed) {
 		super(materialIn);
 		DAMAGE = damage;
 		SPEED = speed;
